backend/requirements.txt
sqlalchemy
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
python-dotenv
pydantic[email]
python-dotenv
